Address TypeSpec helper test review
Move TspClientHelper tests into a dedicated fixture file and guard temporary directory cleanup.


Copilot-Session: 7f90d60d-a51d-43db-a907-5fb0393a763e
